A1 complete (in disaster)
The assignment 1 tic-tac-toe game was not a success. all the UI elements are there. the victory condition has been implemented. However the code is not working even when it should. I have replaced my old code with the MiniMax algorithm but the code goes arye once played. the code should be working but doesn't.

tried to fix them but no luck.